+++ client/src/components/ProductList/styles.less
@@ -1,0 +1,5 @@
+.products-list {
+  display: flex;
+  justify-content: space-between;
+  padding: 50px 0;
+}
